Manage Orders works
After a customer submits an order, it goes through a series of checks:
- The  customer can go back to the order and mark it as "cancelled"
- If the customer hasn't marked the order as "cancelled", the business can accept / reject the order, set the min temp and max temp for the order.
- If the order was accepted by the business, it's displayed to the selected delivery service who can reject or set the order as "delivering", it can also change the delivery person from it's own delivery people.
- If the order was set to "delivering" by the delivery service, it is shown to the selected delivery person who can mark it as completed once it's completed.
Once it's marked as completed, the current time will be set as set as the order's delivery time.

The order can be chatted in the message center until the order has reached "cancelled", "rejected" or "completed" status.